refactor: resource abstractor updated to use unquoted json
