+++ resource-abstractor/resource_abstractor.py
@@ -8,8 +8,7 @@ from flask import Flask
 from flask_cors import CORS
 from flask_smorest import Api
 from flask_swagger_ui import get_swaggerui_blueprint
-from flask.json import JSONEncoder
-from bson import ObjectId
+from utils.json_encoder import MongoJSONEncoder
 
 # Configure logging with environment variable, default to DEBUG
 log_level_str = os.environ.get("LOG_LEVEL", "DEBUG").upper()
@@ -30,13 +29,6 @@ logging.getLogger("pymongo.serverSelection").setLevel(logging.WARNING)
 
 RESOURCE_ABSTRACTOR_PORT = os.environ.get("RESOURCE_ABSTRACTOR_PORT")
 
-
-
-class MongoJSONEncoder(JSONEncoder):
-    def default(self, obj):
-        if isinstance(obj, ObjectId):
-            return str(obj)
-        return super().default(obj)
 
 app = Flask(__name__)
 app.json_encoder = MongoJSONEncoder
